refactor(lint-contrast): migrate to packages/ workspace structure
Move scripts/lint-contrast.ts and scripts/lib/contrast/ to
packages/lint-contrast/ as an independent pnpm workspace package
(@vrchat-albums/lint-contrast). Includes package.json, tsconfig.json,
vitest.config.ts, bin/lint-contrast.ts, src/cli.ts, src/index.ts,
all src/ modules, test fixtures, and split unit tests.

Root package.json: update lint:contrast script to delegate via pnpm
--filter, and remove oxc-parser/culori/@types/culori from root deps.
