+++ package.json
@@ -51,7 +51,7 @@
     "sync-landing-images": "bash scripts/sync-landing-images.sh",
     "lint": "run-p lint:oxlint lint:fmt-check lint:type-check lint:actionlint lint:ast-grep lint:custom",
     "lint:custom": "tsx scripts/lint-custom.ts",
-    "lint:contrast": "tsx scripts/lint-contrast.ts",
+    "lint:contrast": "pnpm --filter @vrchat-albums/lint-contrast run lint-contrast -- --project $INIT_CWD",
     "lint:valueobjects": "tsx scripts/lint-valueobjects.ts",
     "lint:effect": "tsx scripts/lint-effect.ts",
     "lint:node-types": "tsx scripts/lint-node-types.ts",
@@ -149,7 +149,6 @@
     "@testing-library/dom": "~10.4.1",
     "@testing-library/react": "~16.3.2",
     "@tktco/node-actionlint": "~1.6.0",
-    "@types/culori": "^4.0.1",
     "@types/license-checker": "25.0.6",
     "@types/node": "~22.15.35",
     "@types/react": "^19.2.9",
@@ -159,7 +158,6 @@
     "@vitest/coverage-v8": "^4.1.2",
     "@welldone-software/why-did-you-render": "^10.0.1",
     "autoprefixer": "10.4.23",
-    "culori": "^4.0.2",
     "electron": "~39.8.6",
     "electron-builder": "~26.8.2",
     "electron-builder-squirrel-windows": "~26.8.2",
@@ -168,7 +166,6 @@
     "license-checker": "25.0.1",
     "minimatch": "^10.2.3",
     "npm-run-all2": "~8.0.4",
-    "oxc-parser": "^0.126.0",
     "oxfmt": "^0.41.0",
     "oxlint": "^1.56.0",
     "oxlint-tsgolint": "^0.17.1",
