+++ packages/lint-contrast/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@vrchat-albums/lint-contrast",
+  "version": "0.1.0",
+  "license": "MIT",
+  "bin": {
+    "lint-contrast": "./bin/lint-contrast.ts"
+  },
+  "type": "module",
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "scripts": {
+    "lint-contrast": "tsx bin/lint-contrast.ts",
+    "test": "vitest run",
+    "typecheck": "tsgo --noEmit"
+  },
+  "dependencies": {
+    "consola": "~3.4.2",
+    "culori": "^4.0.2",
+    "glob": "^13.0.0",
+    "oxc-parser": "^0.126.0",
+    "pathe": "2.0.3",
+    "postcss": "8.5.6",
+    "tailwindcss": "^4.1.18",
+    "ts-pattern": "5.9.0"
+  },
+  "devDependencies": {
+    "@types/culori": "^4.0.1",
+    "@vitest/coverage-v8": "^4.1.2",
+    "tsx": "^4.21.0",
+    "typescript": "~5.9.3",
+    "vitest": "^4.1.2"
+  }
+}
