+++ packages/lint-contrast/src/types.ts
