+++ packages/lint-contrast/test-fixtures/ng-alpha-composite.tsx
