+++ packages/lint-contrast/test-fixtures/ng-low-contrast-dark.tsx
