+++ packages/lint-contrast/test-fixtures/ok-card-on-background.tsx
