+++ packages/lint-contrast/test-fixtures/skip-no-colors.tsx
