+++ packages/lint-contrast/test-fixtures/warn-dynamic-class.tsx
